requirements.txt
django==3.2
django-environ==0.11.2
djangorestframework==3.14.0
pyjwt==2.8.0

# deploy
gunicorn==20.1.0
psycopg2-binary==2.9.6